tests (Construct Util V2): move test data to separate file
